Ord items list now contain pointers to the Items
so that it can contain subclasses of Item, since the ones contained are
just the pointers. The space reserved for an Item wouldn't be enough for
a subclass' additional members. The space reserved for *a pointer to* an
Item, on the other hand, should be enough for a later casted pointer to
something derived from an Item.